% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{foxGET}
\alias{foxGET}
\title{Make a GET call to Fox Den}
\usage{
foxGET(endpoint, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a GET call to Fox Den
}
