% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{foxPOST}
\alias{foxPOST}
\title{Make a POST call to Fox Den}
\usage{
foxPOST(endpoint, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a POST call to Fox Den
}
