% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_credentials}
\alias{get_credentials}
\title{Fetch credentials from Fox Den credentials file}
\usage{
get_credentials(path = "~/.foxdenCredentials")
}
\value{
A list with elements \code{email} and \code{password}
}
\description{
Fetch credentials from Fox Den credentials file
}
