% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{list_columns}
\alias{list_columns}
\title{List columns and column destinations of a table}
\usage{
list_columns(table, database = "fox_integrated")
}
\arguments{
\item{table}{The table name}

\item{database}{The name of the database}
}
\description{
List columns and column destinations of a table
}
