% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{list_tables}
\alias{list_tables}
\title{List tables in a database}
\usage{
list_tables(database = "fox_integrated")
}
\arguments{
\item{database}{The name of the database}
}
\description{
List tables in a database
}
