% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{download_table}
\alias{download_table}
\title{Download a table}
\usage{
download_table(table, download_to = NULL, database = "fox_integrated")
}
\arguments{
\item{table}{The table name}

\item{download_to}{Local folder to download the table to}

\item{inheritParams}{list_tables}
}
\description{
Download a table
}
