% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fox}
\alias{fox}
\title{Make a REST call to Fox Den}
\usage{
fox(endpoint, method, body = NULL, ...)
}
\arguments{
\item{endpoint}{Server endpoint.}

\item{method}{function which accepts the API endpoint as its first argument}

\item{body}{The body of the request}

\item{...}{named arguments to pass to \code{httr::modify_url}}
}
\description{
Make a REST call to Fox Den
}
