% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fox_login}
\alias{fox_login}
\title{Log in to Fox Den}
\usage{
fox_login(email = NULL, password = NULL)
}
\arguments{
\item{email}{Your email}

\item{password}{Your Fox Den password}
}
\description{
Log in by inputting your username and password or cache your credentials in
a YAML file with the name \code{.foxdenCredentials} in your home (~) directory.
This file should have keys \code{email} and \code{password}.
}
